+++ src/llm-error-pages/utils.js
@@ -1,0 +1,466 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import { getStaticContent, tracingFetch as fetch } from '@adobe/spacecat-shared-utils';
+
+// ============================================================================
+// CONSTANTS
+// ============================================================================
+
+export const LLM_USER_AGENT_PATTERNS = {
+  chatgpt: '(?i)ChatGPT|GPTBot|OAI-SearchBot',
+  perplexity: '(?i)Perplexity',
+  claude: '(?i)Claude|Anthropic',
+  gemini: '(?i)Gemini',
+  copilot: '(?i)Copilot',
+};
+
+const TIME_CONSTANTS = {
+  ISO_MONDAY: 1,
+  ISO_SUNDAY: 0,
+  DAYS_PER_WEEK: 7,
+};
+
+const REGEX_PATTERNS = {
+  URL_SANITIZATION: /[^a-zA-Z0-9]/g,
+  BUCKET_SANITIZATION: /[._]/g,
+};
+
+const CDN_LOGS_PREFIX = 'cdn-logs-';
+
+const VALIDATION_TIMEOUT = 10000; // 10 seconds
+const MAX_CONCURRENT_VALIDATIONS = 20;
+
+// ============================================================================
+// LLM USER AGENT UTILITIES
+// ============================================================================
+
+export function getLlmProviderPattern(provider) {
+  if (typeof provider !== 'string' || !provider.trim()) {
+    return null;
+  }
+  return LLM_USER_AGENT_PATTERNS[provider.toLowerCase()] || null;
+}
+
+export function getAllLlmProviders() {
+  return Object.keys(LLM_USER_AGENT_PATTERNS);
+}
+
+export function buildLlmUserAgentFilter(providers = null) {
+  const targetProviders = providers || getAllLlmProviders();
+  const patterns = targetProviders
+    .map((provider) => getLlmProviderPattern(provider))
+    .filter(Boolean);
+
+  if (patterns.length === 0) {
+    return null;
+  }
+
+  return `REGEXP_LIKE(user_agent, '${patterns.join('|')}')`;
+}
+
+export function normalizeUserAgentToProvider(rawUserAgent) {
+  if (!rawUserAgent || typeof rawUserAgent !== 'string') return 'Unknown';
+
+  if (/chatgpt|gptbot|oai-searchbot/i.test(rawUserAgent)) {
+    return 'ChatGPT';
+  }
+  if (/perplexity/i.test(rawUserAgent)) {
+    return 'Perplexity';
+  }
+  if (/claude|anthropic/i.test(rawUserAgent)) {
+    return 'Claude';
+  }
+  if (/gemini/i.test(rawUserAgent)) {
+    return 'Gemini';
+  }
+  if (/copilot/i.test(rawUserAgent)) {
+    return 'Copilot';
+  }
+
+  return rawUserAgent;
+}
+
+// ============================================================================
+// QUERY BUILDING UTILITIES
+// ============================================================================
+
+function buildDateFilter(startDate, endDate) {
+  const formatPart = (date) => ({
+    year: date.getUTCFullYear().toString(),
+    month: (date.getUTCMonth() + 1).toString().padStart(2, '0'),
+    day: date.getUTCDate().toString().padStart(2, '0'),
+  });
+
+  const start = formatPart(startDate);
+  const end = formatPart(endDate);
+
+  return start.year === end.year && start.month === end.month
+    ? `(year = '${start.year}' AND month = '${start.month}' AND day >= '${start.day}' AND day <= '${end.day}')`
+    : `((year = '${start.year}' AND month = '${start.month}' AND day >= '${start.day}')
+       OR (year = '${end.year}' AND month = '${end.month}' AND day <= '${end.day}'))`;
+}
+
+function buildWhereClause(conditions = [], llmProviders = null, siteFilters = []) {
+  const allConditions = [...conditions];
+
+  // Add LLM user agent filter
+  if (llmProviders) {
+    const llmFilter = buildLlmUserAgentFilter(llmProviders);
+    if (llmFilter) {
+      allConditions.push(llmFilter);
+    }
+  }
+
+  if (siteFilters && siteFilters.length > 0) {
+    allConditions.push(siteFilters);
+  }
+
+  return allConditions.length > 0 ? `WHERE ${allConditions.join(' AND ')}` : '';
+}
+
+export async function buildLlmErrorPagesQuery(options) {
+  const {
+    databaseName,
+    tableName,
+    startDate,
+    endDate,
+    llmProviders = null,
+    siteFilters = [],
+  } = options;
+
+  const conditions = [];
+
+  // Date range filter
+  if (startDate && endDate) {
+    conditions.push(buildDateFilter(startDate, endDate));
+  }
+
+  // Error status filter - get all error status codes (400-599)
+  conditions.push('status BETWEEN 400 AND 599');
+
+  const whereClause = buildWhereClause(conditions, llmProviders, siteFilters);
+
+  return getStaticContent({
+    databaseName,
+    tableName,
+    whereClause,
+  }, './src/llm-error-pages/sql/llm-error-pages.sql');
+}
+
+// ============================================================================
+// SITE AND CONFIGURATION UTILITIES
+// ============================================================================
+
+export function extractCustomerDomain(site) {
+  return new URL(site.getBaseURL()).host
+    .replace(REGEX_PATTERNS.URL_SANITIZATION, '_')
+    .toLowerCase();
+}
+
+export function getAnalysisBucket(customerDomain) {
+  const bucketCustomer = customerDomain.replace(REGEX_PATTERNS.BUCKET_SANITIZATION, '-');
+  return `${CDN_LOGS_PREFIX}${bucketCustomer}`;
+}
+
+export function getS3Config(site) {
+  const customerDomain = extractCustomerDomain(site);
+  const customerName = customerDomain.split(/[._]/)[0];
+  const { bucketName: bucket } = site.getConfig().getCdnLogsConfig()
+    || { bucketName: getAnalysisBucket(customerDomain) };
+
+  return {
+    bucket,
+    customerName,
+    customerDomain,
+    aggregatedLocation: `s3://${bucket}/aggregated/`,
+    databaseName: `cdn_logs_${customerDomain}`,
+    tableName: `aggregated_logs_${customerDomain}`,
+    getAthenaTempLocation: () => `s3://${bucket}/temp/athena-results/`,
+  };
+}
+
+export async function validateDatabaseAndTable(athenaClient, s3Config, log) {
+  const { tableName, databaseName } = s3Config;
+
+  try {
+    const validationQuery = `SELECT 1 FROM ${databaseName}.${tableName} LIMIT 1`;
+    log.info(`Validating database and table: ${databaseName}.${tableName}`);
+
+    await athenaClient.query(validationQuery, databaseName, `[Athena Query] Validate ${databaseName}.${tableName}`);
+    log.info('Database and table validated successfully');
+  } catch (error) {
+    const errorMessage = `Database '${databaseName}' or table '${tableName}' does not exist: ${error.message}`;
+    log.error(errorMessage);
+    throw new Error(errorMessage);
+  }
+}
+
+// ============================================================================
+// DATE AND TIME UTILITIES
+// ============================================================================
+
+export function formatDateString(date) {
+  return date.toISOString().split('T')[0];
+}
+
+function getWeekNumber(date) {
+  const d = new Date(date);
+  d.setUTCHours(0, 0, 0, 0);
+  /* c8 ignore next */
+  d.setUTCDate(d.getUTCDate() + 4 - (d.getUTCDay() || 7));
+  const yearStart = new Date(Date.UTC(d.getUTCFullYear(), 0, 1));
+  return Math.ceil(((d - yearStart) / 86400000 + 1) / 7);
+}
+
+export function getWeekRange(offsetWeeks = 0, referenceDate = new Date()) {
+  const refDate = new Date(referenceDate);
+  const isSunday = refDate.getUTCDay() === TIME_CONSTANTS.ISO_SUNDAY;
+  const daysToMonday = isSunday ? 6 : refDate.getUTCDay() - TIME_CONSTANTS.ISO_MONDAY;
+
+  const weekStart = new Date(refDate);
+  const totalOffset = daysToMonday - (offsetWeeks * TIME_CONSTANTS.DAYS_PER_WEEK);
+  weekStart.setUTCDate(refDate.getUTCDate() - totalOffset);
+  weekStart.setUTCHours(0, 0, 0, 0);
+
+  const weekEnd = new Date(weekStart);
+  weekEnd.setUTCDate(weekStart.getUTCDate() + 6);
+  weekEnd.setUTCHours(23, 59, 59, 999);
+
+  return { weekStart, weekEnd };
+}
+
+export function createDateRange(startInput, endInput) {
+  const startDate = new Date(startInput);
+  const endDate = new Date(endInput);
+
+  if (Number.isNaN(startDate.getTime()) || Number.isNaN(endDate.getTime())) {
+    throw new Error('Invalid date format provided');
+  }
+
+  if (startDate >= endDate) {
+    throw new Error('Start date must be before end date');
+  }
+
+  startDate.setUTCHours(0, 0, 0, 0);
+  endDate.setUTCHours(23, 59, 59, 999);
+
+  return { startDate, endDate };
+}
+
+export function generatePeriodIdentifier(startDate, endDate) {
+  const start = formatDateString(startDate);
+  const end = formatDateString(endDate);
+
+  const diffDays = Math.ceil((endDate - startDate) / (24 * 60 * 60 * 1000));
+  if (diffDays === 7) {
+    const year = startDate.getUTCFullYear();
+    const weekNum = getWeekNumber(startDate);
+    return `w${String(weekNum).padStart(2, '0')}-${year}`;
+  }
+
+  return `${start}_to_${end}`;
+}
+
+export function generateReportingPeriods(referenceDate = new Date()) {
+  const { weekStart, weekEnd } = getWeekRange(-1, referenceDate);
+
+  const weekNumber = getWeekNumber(weekStart);
+  const year = weekStart.getUTCFullYear();
+
+  const weeks = [{
+    weekNumber,
+    year,
+    weekLabel: `Week ${weekNumber}`,
+    startDate: weekStart,
+    endDate: weekEnd,
+    dateRange: {
+      start: formatDateString(weekStart),
+      end: formatDateString(weekEnd),
+    },
+  }];
+
+  return {
+    weeks,
+    referenceDate: referenceDate.toISOString(),
+    columns: [`Week ${weekNumber}`],
+  };
+}
+
+// ============================================================================
+// FILTERING AND PROCESSING UTILITIES
+// ============================================================================
+
+export function buildSiteFilters(filters) {
+  if (!filters || filters.length === 0) return '';
+
+  const clauses = filters.map(({ key, value, type }) => {
+    const regexPattern = value.join('|');
+    if (type === 'exclude') {
+      return `NOT REGEXP_LIKE(${key}, '(?i)(${regexPattern})')`;
+    }
+    return `REGEXP_LIKE(${key}, '(?i)(${regexPattern})')`;
+  });
+
+  const filterConditions = clauses.length > 1 ? clauses.join(' AND ') : clauses[0];
+  return `(${filterConditions})`;
+}
+
+export function processLlmErrorPagesResults(results) {
+  if (!results || results.length === 0) {
+    return {
+      totalErrors: 0,
+      errorPages: [],
+      summary: {
+        uniqueUrls: 0,
+        uniqueUserAgents: 0,
+        statusCodes: {},
+      },
+    };
+  }
+
+  const statusCodes = {};
+  const uniqueUrls = new Set();
+  const uniqueUserAgents = new Set();
+
+  results.forEach((row) => {
+    const status = row.status || 'Unknown';
+    statusCodes[status] = (statusCodes[status] || 0) + (row.total_requests || 0);
+    uniqueUrls.add(row.url);
+    uniqueUserAgents.add(row.user_agent);
+  });
+
+  const totalErrors = Object.values(statusCodes).reduce((sum, count) => sum + count, 0);
+
+  return {
+    totalErrors,
+    errorPages: results,
+    summary: {
+      uniqueUrls: uniqueUrls.size,
+      uniqueUserAgents: uniqueUserAgents.size,
+      statusCodes,
+    },
+  };
+}
+
+// ============================================================================
+// URL VALIDATION UTILITIES
+// ============================================================================
+
+/**
+ * Validates a single URL with simple status consistency check
+ * @param {Object} error - Error object with url, userAgent, rawUserAgents, status
+ * @param {Object} log - Logger instance
+ * @returns {Promise<Object|null>} - Validated error object or null if invalid
+ */
+export async function validateSingleUrl(error, log) {
+  const {
+    url, userAgent, rawUserAgents, status,
+  } = error;
+
+  try {
+    const testUserAgent = rawUserAgents[0]; // Use first raw user agent
+
+    // For 403 errors, check if it's universally blocked
+    if (status === '403') {
+      const simpleResponse = await fetch(url, { timeout: VALIDATION_TIMEOUT });
+      if (simpleResponse.status === 403) {
+        log.debug(`URL ${url} returns 403 even with simple GET - universally blocked, excluding`);
+        return null; // Exclude universal 403s
+      }
+    }
+
+    // Test with LLM user agent to verify status consistency
+    const response = await fetch(url, {
+      method: 'GET',
+      headers: {
+        'User-Agent': testUserAgent,
+      },
+      timeout: VALIDATION_TIMEOUT,
+    });
+
+    // Status consistency check - remove if status changed
+    if (response.status.toString() !== status) {
+      log.debug(`URL ${url} status mismatch - expected ${status}, got ${response.status} - removing`);
+      return null;
+    }
+
+    // URL passed validation
+    log.debug(`URL ${url} validated successfully - ${userAgent} status ${status}`);
+    return {
+      ...error,
+      validatedAt: new Date().toISOString(),
+    };
+  } catch (validationError) {
+    // Network errors, timeouts, etc. - still include in opportunities
+    log.warn(`Validation failed for ${url} (${userAgent}): ${validationError.message} - including anyway`);
+    return {
+      ...error,
+      validatedAt: new Date().toISOString(),
+      validationError: validationError.message,
+    };
+  }
+}
+
+/**
+ * Validates URLs in batches with concurrency control
+ * @param {Array} errors - Array of error objects to validate
+ * @param {Object} log - Logger instance
+ * @returns {Promise<Array>} - Array of validated error objects
+ */
+export async function validateUrlsBatch(errors, log) {
+  log.info(`Starting URL validation for ${errors.length} URLs`);
+
+  const validatedUrls = [];
+
+  // Create batch processing promises to avoid await-in-loop
+  const batchPromises = [];
+  for (let i = 0; i < errors.length; i += MAX_CONCURRENT_VALIDATIONS) {
+    const batch = errors.slice(i, i + MAX_CONCURRENT_VALIDATIONS);
+    const batchIndex = Math.floor(i / MAX_CONCURRENT_VALIDATIONS) + 1;
+    const totalBatches = Math.ceil(errors.length / MAX_CONCURRENT_VALIDATIONS);
+
+    log.info(`Preparing validation batch ${batchIndex}/${totalBatches} (${batch.length} URLs)`);
+
+    // Create promise for this batch
+    const batchPromise = Promise.allSettled(
+      batch.map((error) => validateSingleUrl(error, log)),
+    ).then((batchResults) => ({
+      batchIndex,
+      batch,
+      results: batchResults,
+    }));
+
+    batchPromises.push(batchPromise);
+  }
+
+  // Process all batches
+  const allBatchResults = await Promise.allSettled(batchPromises);
+
+  // Collect valid results from all batches
+  allBatchResults.forEach((batchResult) => {
+    if (batchResult.status === 'fulfilled') {
+      const { batchIndex, results } = batchResult.value;
+      log.info(`Processing results for batch ${batchIndex} (${results.length} URLs)`);
+
+      results.forEach((result) => {
+        if (result.status === 'fulfilled' && result.value !== null) {
+          validatedUrls.push(result.value);
+        }
+      });
+    } else {
+      log.error(`Batch processing failed: ${batchResult.reason}`);
+    }
+  });
+  log.info(`URL validation complete: ${validatedUrls.length}/${errors.length} URLs passed validation`);
+  return validatedUrls;
+}
