+++ test/audits/llm-error-pages/report-utils.test.js
@@ -11,16 +11,24 @@
  */
 
 /* eslint-env mocha */
-import { expect } from 'chai';
+import { expect, use } from 'chai';
 import sinon from 'sinon';
+import sinonChai from 'sinon-chai';
+import {
+  getS3Config, createDateRange, generateReportingPeriods, buildSiteFilters,
+  processLlmErrorPagesResults, validateDatabaseAndTable, generatePeriodIdentifier, getWeekRange,
+} from '../../../src/llm-error-pages/utils.js';
 
 describe('LLM Error Pages - Report Utils', () => {
-  let reportUtils;
   let sandbox;
 
   const mockSite = (baseURL, config = null) => ({
     getBaseURL: () => baseURL,
     getConfig: () => ({ getCdnLogsConfig: () => config }),
+  });
+
+  before(() => {
+    use(sinonChai);
   });
 
   beforeEach(async () => {
@@ -32,13 +40,9 @@ describe('LLM Error Pages - Report Utils', () => {
   });
 
   describe('Direct Import Tests', () => {
-    before(async () => {
-      reportUtils = await import('../../../src/llm-error-pages/utils/report-utils.js');
-    });
-
     describe('getS3Config', () => {
       it('should generate S3 config from site URL', () => {
-        const config = reportUtils.getS3Config(mockSite('https://example.com'));
+        const config = getS3Config(mockSite('https://example.com'));
         expect(config.customerName).to.equal('example');
         expect(config.customerDomain).to.equal('example_com');
         expect(config.databaseName).to.equal('cdn_logs_example_com');
@@ -46,33 +50,33 @@ describe('LLM Error Pages - Report Utils', () => {
       });
 
       it('should handle custom bucket config', () => {
-        const config = reportUtils.getS3Config(mockSite('https://test.com', { bucketName: 'custom-bucket' }));
+        const config = getS3Config(mockSite('https://test.com', { bucketName: 'custom-bucket' }));
         expect(config.bucket).to.equal('custom-bucket');
         expect(config.customerName).to.equal('test');
         expect(config.customerDomain).to.equal('test_com');
       });
 
       it('should handle null config fallback', () => {
-        const config = reportUtils.getS3Config(mockSite('https://empty.com', null));
+        const config = getS3Config(mockSite('https://empty.com', null));
         expect(config.bucket).to.equal('cdn-logs-empty-com');
         expect(config.customerName).to.equal('empty');
         expect(config.customerDomain).to.equal('empty_com');
       });
 
       it('should handle empty object config', () => {
-        const config = reportUtils.getS3Config(mockSite('https://empty.com', {}));
+        const config = getS3Config(mockSite('https://empty.com', {}));
         expect(config.bucket).to.be.undefined;
         expect(config.customerName).to.equal('empty');
         expect(config.customerDomain).to.equal('empty_com');
       });
 
       it('should generate correct temp location', () => {
-        const config = reportUtils.getS3Config(mockSite('https://test.example.com'));
+        const config = getS3Config(mockSite('https://test.example.com'));
         expect(config.getAthenaTempLocation()).to.equal('s3://cdn-logs-test-example-com/temp/athena-results/');
       });
 
       it('should handle complex domain names', () => {
-        const config = reportUtils.getS3Config(mockSite('https://sub-domain.multi-word-site.example-test.co.uk'));
+        const config = getS3Config(mockSite('https://sub-domain.multi-word-site.example-test.co.uk'));
         expect(config.customerDomain).to.equal('sub_domain_multi_word_site_example_test_co_uk');
         expect(config.customerName).to.equal('sub');
       });
@@ -80,7 +84,7 @@ describe('LLM Error Pages - Report Utils', () => {
 
     describe('Date and Time Operations', () => {
       it('should create and validate date ranges', () => {
-        const { startDate, endDate } = reportUtils.createDateRange('2025-01-01', '2025-01-07');
+        const { startDate, endDate } = createDateRange('2025-01-01', '2025-01-07');
         expect(startDate.getUTCHours()).to.equal(0);
         expect(startDate.getUTCMinutes()).to.equal(0);
         expect(startDate.getUTCSeconds()).to.equal(0);
@@ -90,17 +94,17 @@ describe('LLM Error Pages - Report Utils', () => {
       });
 
       it('should throw error for invalid date format', () => {
-        expect(() => reportUtils.createDateRange('invalid', '2025-01-07')).to.throw('Invalid date format provided');
-        expect(() => reportUtils.createDateRange('2025-01-01', 'invalid')).to.throw('Invalid date format provided');
+        expect(() => createDateRange('invalid', '2025-01-07')).to.throw('Invalid date format provided');
+        expect(() => createDateRange('2025-01-01', 'invalid')).to.throw('Invalid date format provided');
       });
 
       it('should throw error when start date is after end date', () => {
-        expect(() => reportUtils.createDateRange('2025-01-07', '2025-01-01')).to.throw('Start date must be before end date');
+        expect(() => createDateRange('2025-01-07', '2025-01-01')).to.throw('Start date must be before end date');
       });
 
       it('should generate valid week ranges', () => {
         [new Date('2025-01-15T10:00:00Z'), new Date('2025-01-07T10:00:00Z')].forEach((date) => {
-          const periods = reportUtils.generateReportingPeriods(date);
+          const periods = generateReportingPeriods(date);
           expect(periods.weeks).to.be.an('array').with.lengthOf(1);
           expect(periods.columns).to.be.an('array');
           expect(periods.referenceDate).to.be.a('string');
@@ -116,7 +120,7 @@ describe('LLM Error Pages - Report Utils', () => {
           new Date('2024-02-29'), // Leap year test
           new Date('2025-01-07T10:00:00Z'),
         ].forEach((date) => {
-          const periods = reportUtils.generateReportingPeriods(date);
+          const periods = generateReportingPeriods(date);
           expect(periods.weeks).to.be.an('array').with.lengthOf(1);
           expect(periods.columns).to.be.an('array');
           expect(periods.referenceDate).to.be.a('string');
@@ -127,31 +131,31 @@ describe('LLM Error Pages - Report Utils', () => {
 
     describe('Site Filters', () => {
       it('should build site filters from array', () => {
-        expect(reportUtils.buildSiteFilters([])).to.equal('');
-        expect(reportUtils.buildSiteFilters(null)).to.equal('');
-        expect(reportUtils.buildSiteFilters([{ key: 'domain', value: ['example.com'] }]))
+        expect(buildSiteFilters([])).to.equal('');
+        expect(buildSiteFilters(null)).to.equal('');
+        expect(buildSiteFilters([{ key: 'domain', value: ['example.com'] }]))
           .to.equal("(REGEXP_LIKE(domain, '(?i)(example.com)'))");
       });
 
       it('should handle multiple values in single filter', () => {
-        expect(reportUtils.buildSiteFilters([{ key: 'domain', value: ['example.com', 'test.com'], type: 'include' }]))
+        expect(buildSiteFilters([{ key: 'domain', value: ['example.com', 'test.com'], type: 'include' }]))
           .to.equal("(REGEXP_LIKE(domain, '(?i)(example.com|test.com)'))");
       });
 
       it('should handle exclude type filters', () => {
-        expect(reportUtils.buildSiteFilters([{ key: 'domain', value: ['example.com', 'test.com'], type: 'exclude' }]))
+        expect(buildSiteFilters([{ key: 'domain', value: ['example.com', 'test.com'], type: 'exclude' }]))
           .to.equal("(NOT REGEXP_LIKE(domain, '(?i)(example.com|test.com)'))");
       });
 
       it('should handle multiple filters', () => {
-        expect(reportUtils.buildSiteFilters([
+        expect(buildSiteFilters([
           { key: 'domain', value: ['example.com'] },
           { key: 'status', value: ['200'] },
         ])).to.equal("(REGEXP_LIKE(domain, '(?i)(example.com)') AND REGEXP_LIKE(status, '(?i)(200)'))");
       });
 
       it('should handle mixed include and exclude filters', () => {
-        expect(reportUtils.buildSiteFilters([
+        expect(buildSiteFilters([
           { key: 'domain', value: ['example.com', 'test.com'], type: 'exclude' },
           { key: 'status', value: ['200'], type: 'include' },
         ])).to.equal("(NOT REGEXP_LIKE(domain, '(?i)(example.com|test.com)') AND REGEXP_LIKE(status, '(?i)(200)'))");
@@ -172,7 +176,7 @@ describe('LLM Error Pages - Report Utils', () => {
           },
         ];
 
-        const processed = reportUtils.processLlmErrorPagesResults(mockResults);
+        const processed = processLlmErrorPagesResults(mockResults);
 
         expect(processed.totalErrors).to.equal(6);
         expect(processed.errorPages).to.have.length(3);
@@ -182,7 +186,7 @@ describe('LLM Error Pages - Report Utils', () => {
       });
 
       it('should handle empty results', () => {
-        const processed = reportUtils.processLlmErrorPagesResults([]);
+        const processed = processLlmErrorPagesResults([]);
 
         expect(processed.totalErrors).to.equal(0);
         expect(processed.errorPages).to.have.length(0);
@@ -192,7 +196,7 @@ describe('LLM Error Pages - Report Utils', () => {
       });
 
       it('should handle null results', () => {
-        const processed = reportUtils.processLlmErrorPagesResults(null);
+        const processed = processLlmErrorPagesResults(null);
 
         expect(processed.totalErrors).to.equal(0);
         expect(processed.errorPages).to.have.length(0);
@@ -224,7 +228,7 @@ describe('LLM Error Pages - Report Utils', () => {
           },
         ];
 
-        const processed = reportUtils.processLlmErrorPagesResults(mockResults);
+        const processed = processLlmErrorPagesResults(mockResults);
 
         expect(processed.totalErrors).to.equal(0);
         expect(processed.errorPages).to.have.length(3);
@@ -247,7 +251,7 @@ describe('LLM Error Pages - Report Utils', () => {
       };
       const mockLog = { info: sandbox.stub(), error: sandbox.stub() };
 
-      await reportUtils.validateDatabaseAndTable(mockAthenaClient, mockS3Config, mockLog);
+      await validateDatabaseAndTable(mockAthenaClient, mockS3Config, mockLog);
 
       expect(mockAthenaClient.query).to.have.been.called;
       expect(mockLog.info).to.have.been.calledWith('Validating database and table: test_database.test_table');
@@ -267,7 +271,7 @@ describe('LLM Error Pages - Report Utils', () => {
       const mockLog = { info: sandbox.stub(), error: sandbox.stub() };
 
       try {
-        await reportUtils.validateDatabaseAndTable(mockAthenaClient, mockS3Config, mockLog);
+        await validateDatabaseAndTable(mockAthenaClient, mockS3Config, mockLog);
         expect.fail('Should have thrown an error');
       } catch (error) {
         expect(error.message).to.include('Database \'nonexistent_db\' or table \'nonexistent_table\' does not exist');
@@ -282,7 +286,7 @@ describe('LLM Error Pages - Report Utils', () => {
       const startDate = new Date('2025-01-06T00:00:00Z'); // Monday
       const endDate = new Date('2025-01-13T00:00:00Z'); // Next Monday (7 days)
 
-      const result = reportUtils.generatePeriodIdentifier(startDate, endDate);
+      const result = generatePeriodIdentifier(startDate, endDate);
 
       // Should be in format w01-2025 (week 1 of 2025)
       expect(result).to.match(/^w\d{2}-\d{4}$/);
@@ -292,7 +296,7 @@ describe('LLM Error Pages - Report Utils', () => {
       const startDate = new Date('2025-01-01T00:00:00Z');
       const endDate = new Date('2025-01-05T00:00:00Z'); // 4 days, not 7
 
-      const result = reportUtils.generatePeriodIdentifier(startDate, endDate);
+      const result = generatePeriodIdentifier(startDate, endDate);
 
       expect(result).to.equal('2025-01-01_to_2025-01-05');
     });
@@ -301,7 +305,7 @@ describe('LLM Error Pages - Report Utils', () => {
       const startDate = new Date('2025-01-01T00:00:00Z');
       const endDate = new Date('2025-01-31T00:00:00Z'); // 30 days
 
-      const result = reportUtils.generatePeriodIdentifier(startDate, endDate);
+      const result = generatePeriodIdentifier(startDate, endDate);
 
       expect(result).to.equal('2025-01-01_to_2025-01-31');
     });
@@ -312,7 +316,7 @@ describe('LLM Error Pages - Report Utils', () => {
       it('should calculate week range correctly for a Sunday reference date', () => {
         const sundayDate = new Date('2024-01-07T12:00:00.000Z'); // Sunday
         expect(sundayDate.getUTCDay()).to.equal(0); // Verify it's Sunday
-        const { weekStart, weekEnd } = reportUtils.getWeekRange(0, sundayDate);
+        const { weekStart, weekEnd } = getWeekRange(0, sundayDate);
 
         expect(weekStart).to.be.a('date');
         expect(weekEnd).to.be.a('date');
@@ -325,7 +329,7 @@ describe('LLM Error Pages - Report Utils', () => {
   describe('generateReportingPeriods', () => {
     it('should generate valid week ranges', () => {
       [new Date('2025-01-15T10:00:00Z'), new Date('2025-01-07T10:00:00Z')].forEach((date) => {
-        const periods = reportUtils.generateReportingPeriods(date);
+        const periods = generateReportingPeriods(date);
         expect(periods.weeks).to.be.an('array').with.lengthOf(1);
         expect(periods.columns).to.be.an('array');
         expect(periods.referenceDate).to.be.a('string');
@@ -341,7 +345,7 @@ describe('LLM Error Pages - Report Utils', () => {
         new Date('2024-02-29'), // Leap year test
         new Date('2025-01-07T10:00:00Z'),
       ].forEach((date) => {
-        const periods = reportUtils.generateReportingPeriods(date);
+        const periods = generateReportingPeriods(date);
         expect(periods.weeks).to.be.an('array').with.lengthOf(1);
         expect(periods.columns).to.be.an('array');
         expect(periods.referenceDate).to.be.a('string');
@@ -355,7 +359,7 @@ describe('LLM Error Pages - Report Utils', () => {
         new Date('2024-12-31T23:59:59Z'),
         new Date('2025-01-15T10:00:00Z'),
       ].forEach((date) => {
-        const periods = reportUtils.generateReportingPeriods(date);
+        const periods = generateReportingPeriods(date);
 
         expect(periods.weeks).to.be.an('array').with.lengthOf(1);
         expect(periods.weeks[0].year).to.be.a('number').greaterThan(2020).lessThan(2030);
@@ -367,18 +371,14 @@ describe('LLM Error Pages - Report Utils', () => {
   });
 
   describe('Edge Cases and Error Handling', () => {
-    before(async () => {
-      reportUtils = await import('../../../src/llm-error-pages/utils/report-utils.js');
-    });
-
     it('should handle extreme date ranges', () => {
       // Very old dates
-      const { startDate: oldStart, endDate: oldEnd } = reportUtils.createDateRange('1900-01-01', '1900-01-02');
+      const { startDate: oldStart, endDate: oldEnd } = createDateRange('1900-01-01', '1900-01-02');
       expect(oldStart.getUTCFullYear()).to.equal(1900);
       expect(oldEnd.getUTCFullYear()).to.equal(1900);
 
       // Far future dates
-      const { startDate: futureStart, endDate: futureEnd } = reportUtils.createDateRange('2100-01-01', '2100-01-02');
+      const { startDate: futureStart, endDate: futureEnd } = createDateRange('2100-01-01', '2100-01-02');
       expect(futureStart.getUTCFullYear()).to.equal(2100);
       expect(futureEnd.getUTCFullYear()).to.equal(2100);
     });
@@ -391,7 +391,7 @@ describe('LLM Error Pages - Report Utils', () => {
         total_requests: i + 1,
       }));
 
-      const processed = reportUtils.processLlmErrorPagesResults(largeResults);
+      const processed = processLlmErrorPagesResults(largeResults);
 
       expect(processed.totalErrors).to.be.a('number');
       expect(processed.errorPages).to.have.length(10000);
@@ -412,7 +412,7 @@ describe('LLM Error Pages - Report Utils', () => {
         },
       ];
 
-      const processed = reportUtils.processLlmErrorPagesResults(specialResults);
+      const processed = processLlmErrorPagesResults(specialResults);
 
       expect(processed.totalErrors).to.equal(4);
       expect(processed.errorPages).to.have.length(3);
