+++ test/audits/llm-error-pages/user-agent-patterns.test.js
@@ -13,31 +13,26 @@
 /* eslint-env mocha */
 import { expect } from 'chai';
 import {
-  LLM_USER_AGENT_PATTERNS,
-  getLlmProviderPattern,
-  getAllLlmProviders,
-  buildLlmUserAgentFilter,
-  normalizeUserAgentToProvider,
-} from '../../../src/llm-error-pages/constants/user-agent-patterns.js';
+  getLlmProviderPattern, getAllLlmProviders, buildLlmUserAgentFilter, normalizeUserAgentToProvider,
+} from '../../../src/llm-error-pages/utils.js';
 
-describe('LLM Error Pages - User Agent Patterns', () => {
-  describe('LLM_USER_AGENT_PATTERNS', () => {
-    it('should contain expected LLM providers', () => {
-      expect(LLM_USER_AGENT_PATTERNS).to.have.property('chatgpt');
-      expect(LLM_USER_AGENT_PATTERNS).to.have.property('perplexity');
-      expect(LLM_USER_AGENT_PATTERNS).to.have.property('claude');
-      expect(LLM_USER_AGENT_PATTERNS).to.have.property('gemini');
-      expect(LLM_USER_AGENT_PATTERNS).to.have.property('copilot');
-    });
+describe('LLM Error Pages – user-agent-patterns', () => {
+  // describe('LLM_USER_AGENT_PATTERNS', () => {
+  //   it('should contain expected LLM providers', () => {
+  //     expect(LLM_USER_AGENT_PATTERNS).to.have.property('chatgpt');
+  //     expect(LLM_USER_AGENT_PATTERNS).to.have.property('perplexity');
+  //     expect(LLM_USER_AGENT_PATTERNS).to.have.property('claude');
+  //     expect(LLM_USER_AGENT_PATTERNS).to.have.property('gemini');
+  //     expect(LLM_USER_AGENT_PATTERNS).to.have.property('copilot');
+  //   });
 
-    it('should have case-insensitive patterns', () => {
-      expect(LLM_USER_AGENT_PATTERNS.chatgpt).to.include('(?i)');
-      expect(LLM_USER_AGENT_PATTERNS.perplexity).to.include('(?i)');
-      expect(LLM_USER_AGENT_PATTERNS.claude).to.include('(?i)');
-      expect(LLM_USER_AGENT_PATTERNS.gemini).to.include('(?i)');
-      expect(LLM_USER_AGENT_PATTERNS.copilot).to.include('(?i)');
-    });
-  });
+  //   it('should have regex patterns for each provider', () => {
+  //     Object.values(LLM_USER_AGENT_PATTERNS).forEach((pattern) => {
+  //       expect(pattern).to.be.a('string');
+  //       expect(pattern).to.include('(?i)');
+  //     });
+  //   });
+  // });
 
   describe('getLlmProviderPattern', () => {
     it('should return pattern for valid provider', () => {
@@ -65,7 +60,7 @@ describe('LLM Error Pages - User Agent Patterns', () => {
   });
 
   describe('getAllLlmProviders', () => {
-    it('should return array of all provider keys', () => {
+    it('should return array of provider names', () => {
       const providers = getAllLlmProviders();
       expect(providers).to.be.an('array');
       expect(providers).to.include('chatgpt');
@@ -73,98 +68,82 @@ describe('LLM Error Pages - User Agent Patterns', () => {
       expect(providers).to.include('claude');
       expect(providers).to.include('gemini');
       expect(providers).to.include('copilot');
-      expect(providers).to.have.length(5);
+    });
+
+    it('should return exactly 5 providers', () => {
+      const providers = getAllLlmProviders();
+      expect(providers).to.have.lengthOf(5);
     });
   });
 
   describe('buildLlmUserAgentFilter', () => {
+    it('should build filter for specific providers', () => {
+      const filter = buildLlmUserAgentFilter(['chatgpt', 'claude']);
+      expect(filter).to.include('ChatGPT|GPTBot|OAI-SearchBot');
+      expect(filter).to.include('Claude|Anthropic');
+      expect(filter).to.include('REGEXP_LIKE(user_agent,');
+      expect(filter).to.not.include('Perplexity');
+    });
+
     it('should build filter for all providers when none specified', () => {
       const filter = buildLlmUserAgentFilter();
-      expect(filter).to.include('REGEXP_LIKE(user_agent');
-      expect(filter).to.include('ChatGPT');
+      expect(filter).to.include('ChatGPT|GPTBot|OAI-SearchBot');
       expect(filter).to.include('Perplexity');
-      expect(filter).to.include('Claude');
+      expect(filter).to.include('Claude|Anthropic');
       expect(filter).to.include('Gemini');
       expect(filter).to.include('Copilot');
     });
 
-    it('should build filter for specific providers', () => {
-      const filter = buildLlmUserAgentFilter(['chatgpt', 'claude']);
-      expect(filter).to.include('REGEXP_LIKE(user_agent');
-      expect(filter).to.include('ChatGPT');
-      expect(filter).to.include('Claude');
-      expect(filter).to.not.include('Perplexity');
+    it('should return null for empty providers array', () => {
+      const filter = buildLlmUserAgentFilter([]);
+      expect(filter).to.be.null;
     });
 
-    it('should return null for empty providers', () => {
-      expect(buildLlmUserAgentFilter([])).to.be.null;
-    });
-
-    it('should return null for invalid providers', () => {
-      expect(buildLlmUserAgentFilter(['invalid', 'nonexistent'])).to.be.null;
-    });
-
-    it('should filter out invalid providers and build filter for valid ones', () => {
+    it('should handle mixed valid/invalid providers', () => {
       const filter = buildLlmUserAgentFilter(['chatgpt', 'invalid', 'claude']);
-      expect(filter).to.include('REGEXP_LIKE(user_agent');
-      expect(filter).to.include('ChatGPT');
-      expect(filter).to.include('Claude');
+      expect(filter).to.include('ChatGPT|GPTBot|OAI-SearchBot');
+      expect(filter).to.include('Claude|Anthropic');
       expect(filter).to.not.include('invalid');
     });
   });
 
   describe('normalizeUserAgentToProvider', () => {
-    it('should handle invalid input', () => {
-      expect(normalizeUserAgentToProvider(null)).to.equal('Unknown');
-      expect(normalizeUserAgentToProvider(undefined)).to.equal('Unknown');
-      expect(normalizeUserAgentToProvider('')).to.equal('Unknown');
-      expect(normalizeUserAgentToProvider(123)).to.equal('Unknown');
+    it('should normalize ChatGPT user agents', () => {
+      expect(normalizeUserAgentToProvider('mozilla chatgpt sdd')).to.equal('ChatGPT');
+      expect(normalizeUserAgentToProvider('GPTBot')).to.equal('ChatGPT');
+      expect(normalizeUserAgentToProvider('OAI-SearchBot')).to.equal('ChatGPT');
     });
 
-    it('should normalize ChatGPT variants', () => {
-      expect(normalizeUserAgentToProvider('ChatGPT-User')).to.equal('ChatGPT');
-      expect(normalizeUserAgentToProvider('GPTBot crawler')).to.equal('ChatGPT');
-      expect(normalizeUserAgentToProvider('OAI-SearchBot/1.0')).to.equal('ChatGPT');
-      expect(normalizeUserAgentToProvider('chatgpt mobile')).to.equal('ChatGPT');
+    it('should normalize Perplexity user agents', () => {
+      expect(normalizeUserAgentToProvider('perplexity bot')).to.equal('Perplexity');
+      expect(normalizeUserAgentToProvider('PERPLEXITY')).to.equal('Perplexity');
     });
 
-    it('should normalize Perplexity variants', () => {
-      expect(normalizeUserAgentToProvider('Perplexity-Bot')).to.equal('Perplexity');
-      expect(normalizeUserAgentToProvider('perplexity.ai')).to.equal('Perplexity');
+    it('should normalize Claude user agents', () => {
+      expect(normalizeUserAgentToProvider('claude bot')).to.equal('Claude');
+      expect(normalizeUserAgentToProvider('anthropic')).to.equal('Claude');
     });
 
-    it('should normalize Claude variants', () => {
-      expect(normalizeUserAgentToProvider('Claude-3')).to.equal('Claude');
-      expect(normalizeUserAgentToProvider('Anthropic AI')).to.equal('Claude');
-      expect(normalizeUserAgentToProvider('claude assistant')).to.equal('Claude');
+    it('should normalize Gemini user agents', () => {
+      expect(normalizeUserAgentToProvider('gemini bot')).to.equal('Gemini');
+      expect(normalizeUserAgentToProvider('GEMINI')).to.equal('Gemini');
     });
 
-    it('should normalize Gemini variants', () => {
-      expect(normalizeUserAgentToProvider('Gemini-Pro')).to.equal('Gemini');
-      expect(normalizeUserAgentToProvider('Google Gemini')).to.equal('Gemini');
-      expect(normalizeUserAgentToProvider('gemini-1.5')).to.equal('Gemini');
-    });
-
-    it('should normalize Copilot variants', () => {
-      expect(normalizeUserAgentToProvider('Microsoft Copilot')).to.equal('Copilot');
-      expect(normalizeUserAgentToProvider('GitHub Copilot')).to.equal('Copilot');
-      expect(normalizeUserAgentToProvider('copilot-web')).to.equal('Copilot');
+    it('should normalize Copilot user agents', () => {
+      expect(normalizeUserAgentToProvider('copilot bot')).to.equal('Copilot');
+      expect(normalizeUserAgentToProvider('COPILOT')).to.equal('Copilot');
     });
 
     it('should return original string for unknown user agents', () => {
-      const unknownAgent = 'Mozilla/5.0 (compatible; CustomBot/1.0)';
-      expect(normalizeUserAgentToProvider(unknownAgent)).to.equal(unknownAgent);
-
-      const anotherUnknown = 'SomeOtherBot/2.0';
-      expect(normalizeUserAgentToProvider(anotherUnknown)).to.equal(anotherUnknown);
+      expect(normalizeUserAgentToProvider('mozilla chrome')).to.equal('mozilla chrome');
+      expect(normalizeUserAgentToProvider('unknown bot')).to.equal('unknown bot');
     });
 
-    it('should handle case insensitive matching', () => {
-      expect(normalizeUserAgentToProvider('CHATGPT')).to.equal('ChatGPT');
-      expect(normalizeUserAgentToProvider('pErPlExItY')).to.equal('Perplexity');
-      expect(normalizeUserAgentToProvider('CLAUDE')).to.equal('Claude');
-      expect(normalizeUserAgentToProvider('GEMINI')).to.equal('Gemini');
-      expect(normalizeUserAgentToProvider('COPILOT')).to.equal('Copilot');
+    it('should handle edge cases', () => {
+      expect(normalizeUserAgentToProvider('')).to.equal('Unknown');
+      expect(normalizeUserAgentToProvider(null)).to.equal('Unknown');
+      expect(normalizeUserAgentToProvider(undefined)).to.equal('Unknown');
+      expect(normalizeUserAgentToProvider(123)).to.equal('Unknown');
     });
   });
 });
